requirements.txt
pandas>=1.5.0
python-dotenv
numpy>=1.21.0
scikit-learn>=1.0.0
xgboost>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
tensorflow>=2.12.0
keras>=2.12.0
ta>=0.10.0
joblib>=1.1.0
tqdm>=4.62.0
pyyaml>=6.0
gymnasium>=0.28.1
stable-baselines3>=2.0.0
torch>=1.13.0
gym>=0.21.0